Completada funcionalidad del submenu gestion de usuarios, añadidos comentarios explicativos en todo el codigo

+++ script.sh
@@ -3,7 +3,9 @@
 #bash-version:5.2.15(1)-release
 #Scrip para la ejecucion de diferentes herramientas.
 #FUNCIONES:
+
 #Control de nombre de usuarios
+#Problema, no muestra nada si el usuario no existe, si hago un echo lo devuelve en el return
 function fun_ctrl_usu(){
 	local nombre=$1
 	while true; do
@@ -16,6 +18,7 @@ function fun_ctrl_usu(){
 		fi
 	done
 }
+
 #Menu Modificacion de usuarios
 function fun_menu_usu_mod(){
 	clear
@@ -23,12 +26,13 @@ function fun_menu_usu_mod(){
 	echo -e "${CIAN}1.${ND} Nombre de usuario."	
 	echo -e "${CIAN}2.${ND} Cambiar Home."
 	echo -e "${CIAN}3.${ND} Fecha de caducidad."
-	echo -e "${CIAN}4.${ND} Añadir o eliminar grupos."	
+	echo -e "${CIAN}4.${ND} Añadir a grupos."	
 	echo -e "${CIAN}5.${ND} Contrasena."	
 	echo -e "${CIAN}6.${ND} Shell."	
 	echo -e "${CIAN}7.${ND} Bloquear."	
 	echo -e "${CIAN}8.${ND} Desbloquear."	
 	echo -e "${ROJO}9.${ND} Volver atras."
+	echo -e "${AZUL}========================${ND}"
 }
 #Menu Gestion de usuarios
 function fun_menu_usu (){
@@ -134,12 +138,14 @@ do
 		;;
 		
 		3) #Ataque de diccionario
-			
 			#Añadir Menu Diccionario
-			echo "ATAQUE DE DICCIONARIO: "
+			fun_gen_menu "ATAQUE DE DICCIONARIO "
 			
 			read -p "Introduzca el HASH: " vhash
+			#Creacion de fichero temporal
 			echo $vhash > temp.txt  #Fichero temporal
+			
+			#Comprobacion del tipo de encriptacion
 			hashid -m $vhash
 			
 			#Control valor variable formato
@@ -165,61 +171,68 @@ do
 		;;
 		
 		4) #Fingerprint
+			#Menu fingerprint
 			fun_gen_menu "Fingerprint"
 			read -p "Especifique el objetivo: " objetivo_nmap
 			clear
 			
-			#Control varible parametros
+			#Llamads a menus
+			fun_gen_menu "Fingerprint"
 			fun_finger
+			
+			#Control varible parametros
 			read -p "Introduzca el algoritmo [ sX, sC, sV  o sn ]: " parametros
 			while [[ "$parametros" != "sX" && "$parametros" != "sC" && "$parametros" != "sV" && "$parametros" != "sn" ]]
 			do
 				echo -e "${ROJO}Opcion incorrecta, introduzca uno de los especificados${ND}"
 				read -p "Introduzca el algoritmo [ sX, sC, sV  o sn ]: " parametros
 			done 
+			
+			#Ejecucion comando nmap
 			sudo nmap -$parametros $objetivo_nmap | egrep -e [0-9][0-9]*/ | cut -d " " -f1,2,4
 			
 			# Confirmacion para avanzar
 			read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla			
 		;;
-		
 		5) #Footprinting			
+			#Bucle while para el submenu 
 			i_exif=0
 			while true
 			do
 				fun_exif
+				#While para eleccion de opciones
 				read -p $'\e[33mElige una opcion\e[0m: ' i_exif
 				while [[ $i_exif -lt 1 && $i_exif -gt 4 ]]
 				do
 					echo "ERROR: OPCION NO CONTEMPLADA"
 					read -p $'\e[33mElige una opcion\e[0m: ' i_exif
 				done
+				#Case para las funcionalidades del menu
 				case $i_exif in
-					1)
+					1)#Metadatos de la ruta actual
 						echo "Mostrando los metadatos de los archivos de la ruta actual:"
-							exiftool *			
+						exiftool *			
 						# Confirmacion para avanzar
 						read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 					;;
-					2)
-					
+					2)#Metadatos de la ruta especifica
 						read -p "ESpecifique la ruta de la cual quiera obterner los metadatos: " ruta_metadatos
 						exiftool $ruta_metadatos/ *			
 						# Confirmacion para avanzar
 						read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 					;;
-					3)
-						read -p "ESpecifique la ruta de la cual quiera obterner los metadatos: " fichero_metadatos
+					3)#Metadatosm de fichero especifico
+						read -p "Especifique el fichero del cual quiera obterner los metadatos: " fichero_metadatos
 						exiftool $fichero_metadatos			
 						# Confirmacion para avanzar
 						read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 					;;
-					4)
+					4)#TheHarvester
 						echo "4"
 						read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 						
 					;;
-					5)
+					5)#Salida del submenu
 						break
 					;;
 					*)
@@ -227,136 +240,170 @@ do
 					sleep 0.5
 					;;
 				esac
-			done
-			# Confirmacion para avanzar
-			read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla			
+			done	
 		;;
 		
 		6) #Gestion de usuarios
+			#While para el submenu de usuarios
 			i_usu=0
 			while true
 			do
 				fun_menu_usu
 				read -p $'\e[33mElige una opcion\e[0m: ' i_usu
 				
+				#Control de opciones
 				while [[ $i_usu -lt 1 && $i_usu -gt 4 ]]
 				do
-					echo ""
+					/dev/null
 				done
 				
 				case $i_usu in
-					1)
-						echo "crea usuario"
+					1)#Creacion de usuario
+						fun_gen_menu "Ceracion de usuario"
+					
 						read -p "Introduzca el nombre de usuario: " nombreusuario
-						read -p "Introduzca la contrasenia: " contrasena
-						contrasena=$(openssl passwd $contrasena)
-						echo $contrasena
-						sudo useradd $nombreusuario --password $contrasena					
-						# Confirmacion para avanzarsudo
+						read -sp "Introduzca la contrasenia: " contrasena
+						echo " "
+						read -sp "Repita la contrasenia: " contrasena2
+						
+						#Comprobacion contrasenia
+						if [ "$contrasena" = "$contrasena2" ]
+						then 
+							contrasena=$(openssl passwd $contrasena)
+							sudo useradd $nombreusuario --password $contrasena					
+						else
+							echo "Las contrasenas no coinciden"
+						fi
+					
+						# Confirmacion para avanzar
 						read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 					;;
-					2)
-						echo "modifica usuario"
+					2)#Modificar usuario
+						fun_gen_menu "modifica usuario"
 						i_usu_mod=0
 						while true
 						do 
 							fun_menu_usu_mod
 							read -p $'\e[33mElige una opcion\e[0m: ' i_usu_mod
-				
-							while [[ $i_usu_mod -lt 1 && $i_usu_mod -gt 4 ]]
-							do
-							echo ""
-							done
 							
 							case $i_usu_mod in
-							1)
+							1) #Cambio de nombre 
 								read -p "Introduzca el nombre del usuario que desea modificar: " nombreviejo
 								
+								#Comprovacion si usuario existe
 								nombreviejo=$(fun_ctrl_usu "$nombreviejo")
 								read -p "Introduzca el nuevo nombre de usuario: " nuevonombre
-								nuevonombre=$(fun_ctrl_usu "$nuevonombre")
+								
 								sudo usermod -l $nuevonombre $nombreviejo
 								echo "NOMBRE DE USUARIO ACTUALIZADO"
 				
-								# Confirmacion para avanzarsudo
+								# Confirmacion para avanzar
 								read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 							;;
-							2)
-								echo -e "${CIAN}2.${ND} Cambiar Home."
-								
+							2)#Cambiar HOME
+								#Comprovacion si usuario existe
 								read -p "Introduzca el nombre del usuario que desea modificar: " nombreusu
 								nombreusu=$(fun_ctrl_usu "$nombreusu")
 								
+								#Comprovacion directorio existe
 								read -p "Introduzca la ruta del nuevo home del usuario: " nuevohome
 								until [[ -n $nuevohome && -d $nuevohome ]]; do
 									read -p "Introduzca la ruta del nuevo home del usuario: " nuevohome
 								done
 								
+								#Cambio
 								sudo usermod --home $nuevohome $nombreusu
-								# Confirmacion para avanzarsudo
+								# Confirmacion para avanzar
 								read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 							;;
-							3)
-								echo -e "${CIAN}3.${ND} Fecha de caducidad."
-								
-								
+							3)#Fecha caducidad
+								#Comprovacion si usuario existe
 								read -p "Introduzca el nombre del usuario que desea modificar: " nombreusu
 								nombreusu=$(fun_ctrl_usu "$nombreusu")
 								
-								#YYYY-MM-DD
-								read -p "Introduzca la nueva fecha de caducidad en formato YYYY-MM-DD: "
-								sudo usermod --expiredate $fechacad $nombreusu
-								# Confirmacion para avanzarsudo
+								#Comprovacion fecha en formato YYYY-MM-DD
+								read -p "Introduzca la nueva fecha de caducidad en formato YYYY-MM-DD: " fechacad
+								if [[ $fechacad =~ ^[0-9]{4}-[0-9]{2}-[0-9]{2}$ ]] && date -d "$fechacad" >/dev/null 2>&1
+								then
+									echo "Modificando usuario: "
+									sudo usermod --expiredate $fechacad $nombreusu
+								else 
+									echo "Fecha no valida o en formato incorrecto"
+								fi
+								#comprobacion con chage -l
+								
+								# Confirmacion para avanzar
 								read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 							;;
-							4)
-								echo -e "${CIAN}4.${ND} Añadir o eliminar grupos."	
-								
+							4)#Añadir a grupo
+								#Comprovacion si usuario existe
 								read -p "Introduzca el nombre del usuario que desea modificar: " nombreusu
 								nombreusu=$(fun_ctrl_usu "$nombreusu")
 								
-								read -p "Introduzca los grupos a los que quiere añadir o eliminar: "
-								sudo usermods $grupos $nombreusu
+								read -p "Introduzca los grupos a los que quiere añadir: " grupos 
+								sudo usermod -a --groups $grupos $nombreusu
 								
-								# Confirmacion para avanzarsudo
+								# Confirmacion para avanzar
 								read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 							;;
-							5)
-								echo -e "${CIAN}5.${ND} Contrasena."
-								
+							5)#Contraseña
+								#Comprovacion si usuario existe
 								read -p "Introduzca el nombre del usuario que desea modificar: " nombreusu
 								nombreusu=$(fun_ctrl_usu "$nombreusu")
 								
+								#Peticion de contraseñas a usuario	
+								read -sp "Introduzca la contrasenia: " contrasena
+								echo " "
+								read -sp "Repita la contrasenia: " contrasena2
+								echo ""
 								
-								# Confirmacion para avanzarsudo
+								#Comprobacion contrasenia
+								if [ "$contrasena" = "$contrasena2" ]
+								then 
+									contrasena=$(openssl passwd $contrasena)
+									sudo usermod --password $contrasena $nombreusu					
+								else
+									echo "Las contrasenas no coinciden"
+								fi
+								# Confirmacion para avanzar
 								read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 							;;
-							6)
-								echo -e "${CIAN}6.${ND} Shell."	
-								
+							6)#Cambio de shell
+
+								#Comprovacion si usuario existe
 								read -p "Introduzca el nombre del usuario que desea modificar: " nombreusu
 								nombreusu=$(fun_ctrl_usu "$nombreusu")
 								
-								# Confirmacion para avanzarsudo
+								#Comprovacion de shell correcta
+								read -p "Introduzca la nueva shell para el usuario [/bin/bash, /bin/sh, /bin/zsh]: " nuevashell
+								while [[ "$nuevashell" != "/bin/bash" && "$nuevashell" != "/bin/sh" && "$nuevashell" != "/bin/zsh" ]]
+								do
+									echo -e "${ROJO}Opcion incorrecta, introduzca una de los especificados${ND}"
+									read -p "Introduzca la nueva shell para el usuario [/bin/bash, /bin/sh, /bin/zsh]: " nuevashell
+								done 
+								
+								sudo usermod --shell $nuevashell $nombreusu 
+								# Confirmacion para avanzar
 								read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 							;;
-							7)
-								echo -e "${CIAN}7.${ND} Bloquear."	
-								
+							7)#Bloquear	
+								#Comprovacion si usuario existe
 								read -p "Introduzca el nombre del usuario que desea modificar: " nombreusu
 								nombreusu=$(fun_ctrl_usu "$nombreusu")
 								
+								sudo usermod --lock $nombreusu
 								
-								# Confirmacion para avanzarsudo
+								# Confirmacion para avanzar
 								read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 							;;
-							8)
-								echo -e "${CIAN}8.${ND} Desbloquear."	
-								
+							8)#Desbloquear	
+								#Comprovacion si usuario existe
 								read -p "Introduzca el nombre del usuario que desea modificar: " nombreusu
 								nombreusu=$(fun_ctrl_usu "$nombreusu")
 								
-								# Confirmacion para avanzarsudo
+								sudo usermod --unlock $nombreusu
+						
+								# Confirmacion para avanzar
 								read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 							;;
 							9)
@@ -369,10 +416,18 @@ do
 							esac
 						done
 					;;
-					3)
-						echo "Elimina usuario"
+					3)#Eliminacion de usuario
+						fun_gen_menu "Eliminar usuario"
 						read -p "Introduzca el nombre de usuario que desea eliminar: " nombreusuario
-						sudo userdel $nombreusuario
+						
+						#Confirmacion de eliminacion de usuario
+						read -p "CONFIRMACION: realmente quiere eliminar el usuario? (s/n)" confirmacion
+						if [ "$confirmacion" = "s" ] || [ "$confirmacion" = "S" ]
+						then 
+							sudo userdel $nombreusuario
+						else
+							echo "Operacion abortada"
+						fi
 						# Confirmacion para avanzar
 						read -rsp $'Pulsa cualquier tecla para continuar...\n' -n1 tecla
 					;;
